requirements.txt
websocket-client
websocket-server
enum34
dronekit
dronekit-sitl
numpy
